discord-sync/pyproject.toml
[tool.poetry]
name = "lwg-replays"
version = "0.1.0"
description = ""
authors = ["bijoo-gh <194651220+bijoo-gh@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
discord-py = "^2.4.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
